fix(state-sync): address PR #112 review comments
Six review-comment fixes:

- form-control-state: re-arm the late-attach retry latch when
  #activeControl identity changes, so a freshly-mounted directive whose
  FormControl registers asynchronously gets its own one-shot retry.
- pending-state.utils.spec: add Signal<DebouncedPendingStateOptions>
  unit tests covering options-change-before-pending and
  options-change-during-pending. Moved here from the acceptance spec
  since they exercise the utility directly.
- state-sync-correctness.spec: drop the duplicated wrapper-level
  pendingDebounce block (now covered as unit tests in pending-state).
- state-sync-correctness.spec: switch the late-attach test to the
  documented one-way [ngModel] + signal-backed model pattern so the
  test doubles as a usage example.
- state-sync-correctness.spec: replace fixed setTimeout(300) waits in
  combos 1 and 4 with a value-mutation + whenStable() round, making
  the negative assertions strictly stronger and faster (731ms → 70ms).
- public-api: re-export DebouncedPendingStateOptionsInput so consumers
  can type Signal-backed configs without a deep import.
- .releaserc + package.json: drop @semantic-release/changelog plugin
  and devDependency to keep the repo internally consistent with the
  GitHub-Releases-only workflow.
