+++ .releaserc
@@ -37,7 +37,6 @@
   ],
   "plugins": [
     "@semantic-release/release-notes-generator",
-    "@semantic-release/changelog",
     [
       "@semantic-release/npm",
       {
